% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visconfig.R
\name{request_vis_config}
\alias{request_vis_config}
\title{Request the Visualization Configuration}
\usage{
request_vis_config(id, session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{id}{The output id of the \code{gwalkr} widget.}
\item{session}{The shiny session object.}
}
\value{
No return value. The result will be available on the client side via
\code{input[[paste0(id, "_visConfig")]]}.
}
\description{
Send a message to the browser to export the current visualization configuration
of a \code{gwalkr} widget.
}
