% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwalkr.R
\name{gwalkr}
\alias{gwalkr}
\title{Create GWalkR Interface in "Viewer"}
\usage{
gwalkr(data, lang = "en")
}
\arguments{
\item{data}{A data frame to be visualized in the GWalkR. The data frame should not be empty.}

\item{lang}{A character string specifying the language for the widget. Possible values are "en" (default), "ja", "zh".}
}
\description{
Use this function to create a GWalkR interface from a given data frame in your "Viewer" window, and start your data exploration! Please make sure the width and the height of your "Viewer" window are large enough.
}
\examples{
\dontrun{
data(mtcars)
gwalkr(mtcars)
}

}
